pyproject.toml
[project]
name = "youtube-stats"
version = "0.1.0"
description = "Analyze your YouTube watch history and generate a personalized YouTube Wrapped"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.100.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
youtube-wrapped = "youtube_wrapped:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
